part2/requirements.txt
boto3
python-dotenv
openai
huggingface_hub
transformers